% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpcuting.R
\name{cut_snp_matrices}
\alias{cut_snp_matrices}
\title{Cut Chromosome SNP Data into Windows}
\usage{
cut_snp_matrices(
  chr_list,
  position_list,
  interval,
  radius = NULL,
  keep_position = FALSE
)
}
\arguments{
\item{chr_list}{A list of matrices, each matrix representing a chromosome's SNP data.}

\item{position_list}{A list of numeric vectors, where each vector indicates the positions of SNPs within the corresponding chromosome.}

\item{interval}{Numeric, the size of each window.}

\item{radius}{Numeric, the radius of each window. If NULL, it defaults to half of the interval size.}

\item{keep_position}{Logical, whether to return the position list for each window (default is FALSE).}
}
\value{
A list of lists of matrices, where each inner list contains matrices representing the windows for a specific chromosome.
}
\description{
This function divides a list of chromosome matrices into smaller windows of a specified size, which is useful for genomic analysis.
}
\examples{
chr_list <- lapply(1:5, function(g) { matrix(rbinom(100 * 150, 2, 0.5), 100) })
position_list <- lapply(1:5, function(g) { runif(150, 0, 100) })
windows <- cut_snp_matrices(chr_list, position_list, interval = 10)
}
