% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_wrapper_functions.R
\name{fast_rhm}
\alias{fast_rhm}
\title{Fast Regional Heritability Mapping Analysis}
\usage{
fast_rhm(y, V_list, D_list)
}
\arguments{
\item{y}{A numeric vector representing the phenotype data.}

\item{V_list}{A list of matrices, where each matrix contains the left singular vectors from the SVD of a genomic data subblock.}

\item{D_list}{A list of numeric vectors, where each vector contains the singular values from the SVD of a genomic data subblock.}
}
\value{
A list with the following components:
\describe{
  \item{\code{R2}}{A matrix containing the R-squared values for each similarity matrix and number of axis considered.}
  \item{\code{R2_weighted}}{A matrix containing the weighted R-squared values for each similarity matrix and number of axis considered.}
}
}
\description{
Performs Regional Heritability Mapping (RHM) analysis on a list of individual similarity matrices. Each similarity matrix is represented by its Singular Value Decomposition (SVD).
}
\details{
The function computes the R-squared values and weighted R-squared values for each combination of similarity matrices and number of axis to consider.
The phenotype data is first scaled and centered before performing the calculations.
}
\examples{
dat = generate_test_data()
result = fast_rhm(dat$y, dat$V_list, dat$D_list)
R2 = result$R2
R2_weighted = result$R2_weighted

}
