% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_snp_data}
\alias{generate_snp_data}
\title{Generate SNP Data for Simulation Studies}
\usage{
generate_snp_data(n_chr = 3, n_rows = 100, n_cols = 100, by_chr = TRUE)
}
\arguments{
\item{n_chr}{Integer, the number of groups of chromosomes to generate. Default is 1.}

\item{n_rows}{Integer, the number of rows in each chromosome matrix. Default is 100.}

\item{n_cols}{Integer, the number of columns in each chromosome matrix. Default is 100.}

\item{by_chr}{Logical, whether to generate data by chromosome (TRUE) or by SNP (FALSE). Default is TRUE.}
}
\value{
A list containing two components:
\itemize{
  \item{chr_list}{List of matrices, where each matrix represents a chromosome.}
  \item{position_list}{List of numeric vectors, where each vector represents the position of each column in the corresponding chromosome matrix.}
}
}
\description{
This function generates simulated SNP data for statistical modeling, particularly useful
in contexts where genetic data are involved. It constructs a hierarchical structure of
chromosome data (chr_list) and position data (position_list) based on the parameters specified.
}
\examples{
snp_data = generate_snp_data()
names(snp_data)
}
