% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate Test Data for Simulation Studies}
\usage{
generate_test_data(
  n = 400,
  k = 3,
  p = 15,
  nu = 5,
  mu = 0,
  signal_percentage = 0.5,
  n_signal = 3,
  contiguous = FALSE
)
}
\arguments{
\item{n}{Integer, the number of observations in each generated matrix. Default is 400}

\item{k}{Integer, the number of groups of matrices to generate. Each group will have `p` matrices. Default is 3.}

\item{p}{Integer, the number of matrices (features) in each group. Default is 15.}

\item{nu}{Integer, the number of columns in each matrix within a group, effectively controlling the dimensionality. Default is 5.}

\item{mu}{Numeric, the global mean used to generate the response vector `y`. Default is 0.}

\item{signal_percentage}{Numeric, the percentage of signal in the data. Default is 0.5.}

\item{n_signal}{Integer, the number of non-zero coefficients in each group. Default is 3.}

\item{contiguous}{Logical, whether to generate contiguous signals in the matrices. Default is FALSE.}
}
\value{
A list containing several components:
\itemize{
  \item{y}{Numeric vector, the response variable generated from the simulated data.}
  \item{V_list}{List of lists, where each sublist contains the left singular vectors of the matrices from X_list.}
  \item{D_list}{List of lists, where each sublist contains the normalized squared singular values of the matrices from X_list.}
  \item{sigma2}{Numeric, the variance used to generate the response vector `y`.}
  \item{mu}{Numeric, the global mean used in the generation of `y`.}
  \item{lambda}{List of numeric vectors, the random effects coefficients used for each group.}
  \item{X_list}{List of lists, where each sublist contains the scaled matrices generated from the data.}
  \item{Z_list}{List of lists, where each sublist contains the random effects used to generate the response vector `y`.}
}
}
\description{
This function generates simulated test data for statistical modeling, particularly useful
in contexts where grouped and multilevel data structures are involved. It constructs
a hierarchical structure of scaled matrix data (X_list), singular vectors (V_list),
normalized singular values (D_list), random effects coefficients (lambda), and
responses (y) based on the parameters specified.
}
\examples{
test_data = generate_test_data()
names(test_data)
}
