% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpcuting.R
\name{get_snp_similarity_matrices}
\alias{get_snp_similarity_matrices}
\title{Calculate Similarity Matrices for SNP Data Windows}
\usage{
get_snp_similarity_matrices(X_list, svd_inertia = 0.95, method = "simple")
}
\arguments{
\item{X_list}{A list of lists of matrices, each representing a window of SNP data.}

\item{svd_inertia}{Numeric, the proportion of inertia (variance) of the SVD to retain (default is 0.95).}

\item{method}{Character or function, the method used to scale the data before computing the Gram matrix. Options are "simple", "van_raden", "yang", or a custom function.}
}
\value{
A list of lists, where each inner list contains pairs of matrices (V and D) representing the SVD decomposition of the similarity matrix for each window.
}
\description{
This function computes the similarity matrices for a list of SNP data windows using specified scaling methods, returning the SVD decomposition for reduced size and increased stability.
}
\examples{
chr_list <- lapply(1:5, function(g) { matrix(rbinom(100 * 150, 2, 0.5), 100) })
position_list <- lapply(1:5, function(g) { runif(150, 0, 100) })
windows <- cut_snp_matrices(chr_list, position_list, interval = 10)
similarity_matrices <- get_snp_similarity_matrices(windows, svd_inertia = 0.95, method = "simple")
}
