% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_wrapper_functions.R
\name{rhm_fusion}
\alias{rhm_fusion}
\title{Regional Heritability Mapping via Gibbs Sampling with a Fusion Regularization Technique}
\usage{
rhm_fusion(
  y,
  V_list,
  D_list,
  niter = 5000,
  thinin = 5,
  burnin = 2000,
  a = 1,
  b = 1,
  alpha = 0.05,
  save_Z = FALSE
)
}
\arguments{
\item{y}{Numeric vector, the response variable from the study.}

\item{V_list}{List of matrices, containing the left singular vectors of the covariance matrices.}

\item{D_list}{List of numeric vectors, containing the normalized singular values of the covariance matrices.}

\item{niter}{Integer, total number of Gibbs sampling iterations to perform.}

\item{thinin}{Integer, thinning interval to reduce autocorrelation in the sampled results.}

\item{burnin}{Integer, number of initial samples to discard to allow the chain to reach stability.}

\item{a}{Numeric, hyperparameter for the prior distribution of the residual variance.}

\item{b}{Numeric, hyperparameter for the prior distribution of the residual variance.}

\item{alpha}{Numeric, the significance level for the selection of random effects coefficients.}

\item{save_Z}{Logical, if TRUE, the function will save the random effects coefficients.}
}
\value{
A list containing sampled values post burn-in for:
\itemize{
  \item{\code{lambda}}{Matrix of random effects coefficients for each group and each iteration.}
  \item{\code{mu}}{Global mean effects across iterations.}
  \item{\code{sigma2}}{Global variance across iterations.}
  \item{\code{tau2}}{Global control of random effects across iterations.}
  \item{\code{Omega2}}{Local control of consecutive differences across iterations.}
  \item{\code{W2}}{Local control of random effects across iterations.}
  \item{\code{Z}}{Random effects across iterations.}
  \item{\code{lambda2_median}}{List of numeric vectors, containing the median of random effects coefficients for each group.}
}
}
\description{
This function performs Gibbs sampling for Bayesian estimation of random effects, using a horseshoe fusion regularization approach.
It handles the random effects by utilizing singular value decomposition components of given covariance matrices derived from genomic sub-blocks.
}
\examples{
dat = generate_test_data()
rhm_fusion(dat$y, dat$V_list, dat$D_list, niter = 2, thinin = 1, burnin = 0)

}
