% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_simulation_functions.R
\name{simulate_mu}
\alias{simulate_mu}
\title{Simulate Global Mean}
\usage{
simulate_mu(sigma2, residuals)
}
\arguments{
\item{sigma2}{Residual variance.}

\item{residuals}{Residuals.}

\item{y}{Response vector.}

\item{Z}{List of matrices representing random effects.}

\item{lambda}{List of coefficients.}

\item{mu}{Mean value.}

\item{V_list}{List of matrices for variance calculations.}
}
\value{
A simulated mean value.
}
\description{
Simulate global mean following the conditional complete posterior distribution.
}
\keyword{internal}
