% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_simulation_functions.R
\name{simulate_Omega2}
\alias{simulate_Omega2}
\title{Simulate Local Control of Differences}
\usage{
simulate_Omega2(lambda, sigma2, tauO2, xi_Omega, Omega2)
}
\arguments{
\item{lambda}{List of coefficients.}

\item{sigma2}{Residual variance.}

\item{tauO2}{Global shrinkage parameter of differences.}

\item{xi_Omega}{Control variable of the Horseshoe prior.}

\item{Omega2}{List of local shrinkage parameters for differences.}
}
\value{
A list of updated local control variables for differences.
}
\description{
Simulate local control of differences following the conditional complete posterior distribution (fused version).
}
\keyword{internal}
