% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_simulation_functions.R
\name{simulate_sigma2_fused}
\alias{simulate_sigma2_fused}
\title{Simulate Residual Variance (Fused)}
\usage{
simulate_sigma2_fused(
  y,
  Z,
  lambda,
  mu,
  tauO2,
  W2,
  Omega2,
  a,
  b,
  V_list,
  q,
  residuals
)
}
\arguments{
\item{y}{Response vector.}

\item{Z}{List of matrices representing random effects.}

\item{lambda}{List of coefficients.}

\item{mu}{Mean value.}

\item{tauO2}{Global shrinkage parameter for differences.}

\item{W2}{List of local shrinkage parameters.}

\item{Omega2}{List of local shrinkage parameters for differences.}

\item{a}{Shape parameter for the prior distribution.}

\item{b}{Rate parameter for the prior distribution.}

\item{V_list}{List of matrices for variance calculations.}

\item{q}{Degrees of freedom.}

\item{residuals}{Residuals.}
}
\value{
A simulated residual variance value.
}
\description{
Simulate residual variance following the conditional complete posterior distribution (fused version).
}
\keyword{internal}
