% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_simulation_functions.R
\name{simulate_tau2_horseshoe}
\alias{simulate_tau2_horseshoe}
\title{Simulate Global Control of Coefficients (Horseshoe)}
\usage{
simulate_tau2_horseshoe(lambda, sigma2, W2, xi_tau, q)
}
\arguments{
\item{lambda}{List of coefficients.}

\item{sigma2}{Residual variance.}

\item{W2}{List of local variance parameters.}

\item{xi_tau}{Control variable of the Horseshoe prior.}

\item{q}{Degrees of freedom.}
}
\value{
A simulated global control variable.
}
\description{
Simulate global control of coefficients following the conditional complete posterior distribution (fused version).
}
\keyword{internal}
