% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_simulation_functions.R
\name{simulate_tauO2_fused}
\alias{simulate_tauO2_fused}
\title{Simulate global shrinkage parameter of differences (Fused)}
\usage{
simulate_tauO2_fused(lambda, sigma2, Omega2, xi_tauO, q)
}
\arguments{
\item{lambda}{List of coefficients.}

\item{sigma2}{Residual variance.}

\item{Omega2}{List of local shrinkage parameters for differences.}

\item{xi_tauO}{Control variable of the Horseshoe prior.}

\item{q}{Degrees of freedom.}
}
\value{
A simulated global control variable.
}
\description{
Simulate global shrinkage parameters of difference following the conditional complete posterior distribution (fused version).
}
\keyword{internal}
