% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_simulation_functions.R
\name{simulate_W2_fusion}
\alias{simulate_W2_fusion}
\title{Simulate Local Control of Coefficients}
\usage{
simulate_W2_fusion(lambda, sigma2, xi_W, W2)
}
\arguments{
\item{lambda}{List of coefficients.}

\item{sigma2}{Residual variance.}

\item{xi_W}{Control variable of the Horseshoe prior.}

\item{W2}{List of local variance parameters.}
}
\value{
A list of updated local control variables for coefficients.
}
\description{
Simulate local control of coefficients following the conditional complete posterior distribution (fused version).
}
\keyword{internal}
