% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhm_simulation_functions.R
\name{simulate_xi}
\alias{simulate_xi}
\title{Simulate Control Variable of Horseshoe Prior}
\usage{
simulate_xi(x2)
}
\arguments{
\item{x2}{Squared coefficients.}
}
\value{
A simulated control variable.
}
\description{
Simulate control variable of the Horseshoe prior.
}
\keyword{internal}
