% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_Z_lambda_fused_cpp}
\alias{simulate_Z_lambda_fused_cpp}
\title{Simulate Random Effects and Coefficients using Rcpp and Armadillo}
\usage{
simulate_Z_lambda_fused_cpp(
  y,
  Z,
  mu,
  lambda,
  W2,
  Omega2,
  sigma2,
  tauO2,
  V_list,
  D_list,
  residuals
)
}
\arguments{
\item{y}{A numeric vector representing the response variable.}

\item{Z}{A list of lists of matrices representing random effects.}

\item{mu}{A double representing the mean value.}

\item{lambda}{A list of vectors representing coefficients.}

\item{W2}{A list of vectors representing the local control of random effects.}

\item{Omega2}{A list of vectors representing the local control of random effects differences.}

\item{sigma2}{A double representing the residual variance.}

\item{tauO2}{A double representing the global control of random effects differences.}

\item{V_list}{A list of lists of left singular matrices used for variance calculations.}

\item{D_list}{A list of lists of matrices for the prior distribution.}

\item{residuals}{A numeric vector representing the residuals.}
}
\value{
An integer value.
}
\description{
This function simulates random effects and coefficients following the conditional complete posterior distribution using Rcpp and Armadillo for performance improvements.
}
